Add a load option for gecko data from St. Olaf radar
The St. Olaf HF radar uses a software package called Gecko developed by
R. Pettersson. The data are saved as .gtd and .gps files. Write a load
script for this file type. As of now, only one file at a time. Future
developments will allow many files to be loaded simultaneously and
concatenated on the fly.