SeisUnix Migration is working, at least for sumigtk
Implement the SU migration routines by saving as .sgy file, executing
the migration command in the terminal, then reading from the resulting
file back into ImpDAR format. This is all done under the hood, so as
long as the user has seisunix installed they should be able to do this
by choosing the 'su' mtype option in impproc migrate.