A big set of backwards-incompatible changes
This is the first step in two parts of breaking changes in order to
streamline files and syntax. I'm using pylint some as a I do this as
well. The goal is to declutter now so that we can advertise at IGS
radioglaciology and not have things break for people between then and
(hopefully) publication.