Sequentialize to reduce peak memory (#18)
* Sequentialize to reduce peak memory

Operations from the command line are now mostly sequential. This
should reduce peak memory usage for most operations. Exceptions are
concatenating and loading gecko or osu files.

* Make the list copy python 2/3 compatible

* Fully test processing, with more options to concatenate profiles with differing picks

* Move Pulse EKKO partitioning into the load() function

As per issue #16, the PE project files were inconsistently handled
because the partitioning was being done in load_and_exit(). Move that
part of the function into load().

Also, change the loop so that the load() function can handle a mix of
project files and standard PE files.

Co-authored-by: Benjamin Hills <bhills@uw.edu>