Auto Picker  (#19)
* First commit for DeLoRES reader

Still needs more work for GPS info. The NMEA strings written on these
files don't seem to be complete.

* Slight change to Pulse Ekko reader for OSU group.

* Change for Pulse Ekko GPS files with more than only GGA strings.

* Slight change for the GSSI reader for data from the OSU group.

* Add a reader for the OSU HF data

* A few small changes for OSU data, interpolation, plotting

* Add a reader for MALA RAMAC data

* Allow alternative data variable names for Olaf Matlab files.

* Update and test all of the analysis scripts.

This is to finalize the scripts for Hills et al. (2020) to be submitted
to the Radioglaciology Annals volume.
A total of 6 attenuation methods implemented:
Method 2 -- Jacobel et al. (2009)
Method 3 -- Schroeder et al. (2016a, 2016b)
Method 5 -- MacGregor et al. (2014); Matsuoka et al. (2010)
Method 6a -- Depth-resolved (regression similar to method 1)
Method 6b -- Depth-resolved (regression similar to method 3)
Method 7 -- Secondary reflection as in Christianson et al. (2016)

* Change analysis file names.

* Update error in attenuation calculations.

* Update load script for ApRES

Accommodate multiple files and multiple bursts

* Add a range differencing function for ApRES.

Following Brennan et al. (2013) and Nicholls et al. (2015) calculate the
vertical motion using a moving window and correlation coefficient
calculated at subsequent depths in two acquisition profiles. Vertical
ice motion between acquisitions can be seen for both internal
reflections and the bed reflector.

* Add a range differencing function to the ApRES processor.

Uses the correlation coefficient between two ApRES profiles to determine
the relative motion between acquisitions.

* Add option for external depth array in ApRES range differencing.

* Add a warning on interpolation

If the gps data onto which the interpolation will be done has longitude
notation different than the radar longitude, then raise an error. We
were having problems with one being written from 0 - 360 and the other
from -180 - 180.

* Some small fixes for typos in hcrop

* Add error to ApRES coherence calculation.

* Update the ApRES range function to output uncertainty

* Small change in ApRES range differencing function

Center the depth profile with each point now representing the center of
the window rather than the top.

* Some changes to Joshua's spectral density plotting.

There were too many options in the plot command which made it confusing.
I cut some out and hopefully made the help file more clear as to how one
could plot the spectral density for each trace using Joshua's function.

* Update specdense plotting to satisfy Travis.

* Add a function for calculating the phase/range uncertainty.

Following Kingslake et al. (2014) calculate uncertainty with a noise
phasor that has random phase and magnitude equal to the median of the
measured phasor.

* Typo in Pulse EKKO load script.

Line to load the time step from the header file was looking for the time
window as 'WINDOW' when a later line has 'AMPLITUDE WINDOW' so it was
loading that number instead.

* Add a load script for the St Olaf matlab files.

These files are mostly the same as the ImpDAR matlab files but commonly
have the data named as 'filtdata', 'interp_data', 'migdata', etc.

* Changes to naming convention for ApRES uncertainty

There were too many variables named similar things, so I made it a bit
more clear.

* Changes to the crop function for pretrig

* Change some notation in the ApRES scripts.

* Typo in Kirchhoff Migration

Removed 1e6 on the depth vector

* Add an autopicking routine.

This routine asks for some picked points on the left side of the image
and will spit out picks across the image for each.

* Allow auto picks from any point on the image

* Fixed some bugs with the sequence of the autopicker.

The autopicker was not working nicely with the normal picker. That is,
when switching back and forth, the program would break. This was because
of an indexing issue with the pick number. Fixed.

* Some of the CReSIS data products are saved as an alternative version of .mat files, allow these to be loaded into ImpDAR

* Add a try/except line for h5py import so that Travis will be happy.

* Created Issue #20

This is in regard to the adaptive horizontal filter. The scaling
function that is used to keep the filter from changing the deeper data
is much more complicated than what I would have used. Considering
simplifying.

* Simplify the scaling function for ahfilt

This was overly complicated and different from that being used in hfilt.
I just copied from hfilt, and cleaned up some of the code. I also added
optionality for setting the window_size from the command line function.

* Change the adaptive hfilt test to make Travis happy.