Update Kirchhoff migration to match distance vectors
At some point distance was changed to be stored in km by default. This
was changed in the c-code migration but not in the python version and
was not compatible with some jupyter tutorial.

The two migration routines are slightly different and I am still trying
to figure out why.