Change NMO constant depth spacing to False (#32)
* Fixes for GSSI for VERY corrupt data

* Update for newer gdal

* Wider errors on vbp test for newer numpy

* Allow no power or time when resampling

* Move to main as the primary branch

* Fix tnum/snum typo in migration printing, stdout flushing

* Allow no power or time when resampling

* Add optionality to load with arbitrary data name

* Add an optionality to gps for gaps in data

The HD 19-20 season had many different GPS issues. In order to process around them I had to change the script here a bit.

* Some small changes to load St Olaf Matlab files.

* Add a load file for TEK data from Catania, Conway, et al.

* Typo in datetime matlab compatibility

* Rearrange arguments in the exec subparsers

Accidentally moved an argument out of processing that needed to be
there.

* Define the decday_interp variable

Test was failing because this variable was not defined in one of the
conditionals

* Change the burst load error so it doesn't fail completely

* Add seconds to the self.decday so that fractional days are preserved.

* Noise phasor below bed

Index the noise phasor from samples below the bed reflection only.

* Update the Apres saving script

Include functionality for saving quad-pol data
Tested.

* Change data type to complex on the range calculation

The data type is float when loaded, but pulse compression should give a
complex array. Information was being lost when the data was saved into a
.mat file, so I changed the datatype on range calculation.

* Small change in load_apres

Keep the shape of apres_data.data consistent no matter how many files
are loaded.

* Update ApRES flags

* Fix typo on data type

* Fix bounds error for nmo depth optimization

* Try capping scipy at v1.8

* Test script for load_tek

* Add a load option for the dual-channel DeepIceIPR

* Fix a longitude decimal minutes conversion

* Load BSI accomodate for test load script which uses the original file structure

* second BSI load test needed a variable definition

* PulseEKKO load timestamp into data with no GPS

* BSI load testing with XIPR dataset

* Change default nmo correction to linear travel time instead of linear depth

* Update NMO function for constant sample spacing

The nmo filter was turning time to nonlinear.
Probably best to keep travel_time as the standard array.
Updated and did some general cleanup to make nmo simpler, especially
with variable velocity.

* One small unit test bug

Co-authored-by: David Lilien <david.lilien@umanitoba.ca>